refactor(core): split checks/fixers into src/base vs src/languages/js (#281)
Pure, output-neutral relocation — proven byte-identical against
pre-refactor snapshots of runDoctor() (45 checks) and getFixers()
(52 fixers).

- src/base/types.ts: shared CheckResult/CheckStatus (re-exported from
  doctor.ts for existing importers).
- src/base/checks.ts: the language-agnostic checks (editorconfig,
  CODEOWNERS, community-health).
- src/languages/js/checks.ts: all Node/pnpm-coupled doctor checks +
  FILE_CHECKS + package/pkg helpers.
- src/languages/js/fixers.ts: the Fixer contract + FIXERS array + fixer
  helpers, with generator imports moving alongside.
- doctor.ts / fix.ts keep their orchestration and now import the moved
  code; runDoctor preserves its exact push order so output is unchanged.

CI/security/repo-hygiene checks (dependabot, codeql, github-settings,
gitlab-ci, coverage, lockfile, AI setup) stay in doctor.ts — issue #282
moves those into src/base next.

Closes #281
